misc changes. Will be abandoned (likely) after next commit

+++ src/main/java/org/pulcifer/climate/svc/diffCalc/Difference.java
@@ -2,6 +2,8 @@ package org.pulcifer.climate.svc.diffCalc;
 
 import org.pulcifer.climate.dto.SimilarCity;
 
+import java.math.BigDecimal;
+
 public interface Difference {
-    float calcDiff(SimilarCity city, SimilarCity similarCity);
+    BigDecimal calcDiff(SimilarCity city, SimilarCity similarCity);
 }
